draft to check if data's saved, if not, get and plot
Script to check if a path for storing data exists, if the desired time range is downloaded, and if not, gets and saves data from IRIS, then plots RSAM, Spectrogram, PPSDs, and Temporal Plots for each day within the specified time range. Currently needs work on the download to PPSD plot part--not linking time segments as a continuous range.