getData as a function + changed logic of checking existing files
Edited from PPSD.ipynb's get_iris_data function, made it so that it checks if files are already saved locally and if not downloads them. Also changed the logic of checking to checking if file name pattern exists or not (ZE.2301..GPZ_YYYYMMDDhhmmss.mseed) as opposed to being based on the time (start_date, end_date). Based on time may have introduced instances of excluding downloading data boarding sections of already downloaded data (this logic was previously performed in the process_data function instead of get_iris_data).