PPSD + temporal plot as a function
Edited to run independently from the process_data function, added the input value options here for the temporal plot. Though, this file by itself may not be useful without running the process_data function, as that displays the spectra to be able to see what frequency bands to use as input values in the temporal plot.